+++ biz/src/main/java/com/hsh24/dms/receipt/dao/IReceiptLogDao.java
@@ -14,6 +14,6 @@ public interface IReceiptLogDao {
 	 * @param receiptLog
 	 * @return
 	 */
-	Long createReceiptLog(ReceiptLog receiptLog);
+	int createReceiptLog(ReceiptLog receiptLog);
 
 }
