+++ biz/src/main/java/com/hsh24/dms/trade/dao/IOrderDao.java
@@ -16,7 +16,7 @@ public interface IOrderDao {
 	 * @param order
 	 * @return
 	 */
-	Long createOrder4Item(Order order);
+	int createOrder4Item(Order order);
 
 	/**
 	 *
