+++ biz/src/main/java/com/hsh24/dms/trade/dao/ITradeLogDao.java
@@ -14,6 +14,6 @@ public interface ITradeLogDao {
 	 * @param tradeLog
 	 * @return
 	 */
-	Long createTradeLog(TradeLog tradeLog);
+	int createTradeLog(TradeLog tradeLog);
 
 }
